back/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
pydantic==2.5.0
httpx==0.25.2
bcrypt==4.1.2
python-jose==3.3.0
passlib==1.7.4
argon2-cffi==25.1.0
